Actualización final
Diseño responsive

+++ src/user/user.service.ts
@@ -14,17 +14,17 @@ export class UserService {
     private readonly jwtService: JwtService,
   ) {}
 
-  // 1. Crear Usuario (con Hash de password)
+  
   async create(createUserDto: CreateUserDto): Promise<User> {
     const { email, password, ...rest } = createUserDto;
 
-    // Verificar si el email ya existe
+    
     const existingUser = await this.userRepository.findOneBy({ email });
     if (existingUser) {
       throw new ConflictException('El correo electrónico ya está registrado');
     }
 
-    // Hashear la contraseña
+    
     const hashedPassword = await bcrypt.hash(password, 10);
 
     const newUser = this.userRepository.create({
@@ -36,12 +36,14 @@ export class UserService {
     return await this.userRepository.save(newUser);
   }
 
-  // 2. Obtener todos los usuarios
+  
   async findAll(): Promise<User[]> {
-    return await this.userRepository.find();
+    return await this.userRepository.find({
+      relations: ['averiasReportadas'],
+    });
   }
 
-  // 3. Buscar por Correo
+  
   async findOne(email: string): Promise<User> {
     const user = await this.userRepository.findOneBy({ email });
     if (!user) {
@@ -50,7 +52,7 @@ export class UserService {
     return user;
   }
 
-  // 4. Actualizar Usuario
+  
   async update(id: number, updateUserDto: Partial<CreateUserDto>): Promise<User> {
     const user = await this.userRepository.preload({
       id: id,
@@ -61,7 +63,7 @@ export class UserService {
       throw new NotFoundException(`Usuario con ID ${id} no existe`);
     }
 
-    // Si se actualiza el password, hay que re-hashearlo
+    
     if (updateUserDto.password) {
       user.password = await bcrypt.hash(updateUserDto.password, 10);
     }
@@ -69,15 +71,19 @@ export class UserService {
     return await this.userRepository.save(user);
   }
 
-  // 5. Eliminar Usuario
+  
   async remove(id: number): Promise<void> {
+    const user = await this.userRepository.findOneBy({ id });
+    if (user && user.email === 'invitado@sistema.com') {
+      throw new ConflictException('No se puede eliminar el usuario de sistema (Sin Registrar)');
+    }
     const result = await this.userRepository.delete(id);
     if (result.affected === 0) {
       throw new NotFoundException(`Usuario con ID ${id} no encontrado`);
     }
   }
 
-  // 6. Verificación de Inicio de Sesión (Login)
+  
   async login(email: string, pass: string) {
     const user = await this.userRepository.findOneBy({ email });
     
@@ -85,14 +91,14 @@ export class UserService {
       throw new UnauthorizedException('Credenciales inválidas');
     }
 
-    // Comparar contraseña enviada con el hash de la DB
+    
     const isMatch = await bcrypt.compare(pass, user.password!);
     
     if (!isMatch) {
       throw new UnauthorizedException('Credenciales inválidas');
     }
 
-    // Generar el JWT
+    
     const payload = { sub: user.id, email: user.email, rol: user.rol };
     
     return {
